libs/orion/requirements.txt
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
requests==2.26.0
urllib3==1.26.6
ratelimit==2.2.1
